requirements.txt
kafka-python==2.0.2
python-binance==1.0.14
requests==2.26.0
